Extract common row accessor methods to GpuColumnVectorBase
